[ME-1591] Cloud Install Wizard for AWS

+++ internal/api/models/token.go
@@ -1,7 +1,7 @@
 package models
 
 type Token struct {
-	ExpiresAt int    `json:"expires_at,omitempty"`
+	ExpiresAt int64  `json:"expires_at,omitempty"`
 	Name      string `json:"name,omitempty"`
 	Role      string `json:"role,omitempty"`
 	Token     string `json:"token,omitempty"`
